Add BTREE_NOATOMICS option
Add the compiler define BTREE_NOATOMIC options to disable using
an atomic reference counter. Instead a normal int will be used
in its place.

This means that using btree_clone() will no longer result in a
clone that is thread-safe, but it will still a valid clone for
single threaded programs. This may be desired (or needed) for
environments where there are no atomics, or where concurrency is
coroutine-based.

Usage:
    cc -DBTREE_NOATOMICS btree.c