Add custom searcher
Using the btree_set_searcher() function, it's now possible to add
a custom search operation on node items during tree traversal.

This is an advanced operation that puts the programmer in control
of how items are searched, such as using a bsearch, sequentially
scanning items, or maybe using SIMD.

The benchmarks now use a custom searcher that sequentially scans
the items instead of a traditional bsearch. Look at the
tests/bench.c file for an example of usage. To bench using a
bsearch use `BSEARCH=1 tests/run.sh bench`.